Refactor the mock data
