+++ processor/test/operation.service.spec.ts
@@ -18,8 +18,7 @@ import { HealthCheckResult } from '@commercetools/connect-payments-sdk/dist/api/
 import * as Config from '../src/config/config';
 
 import { DefaultPaymentService } from '@commercetools/connect-payments-sdk/dist/commercetools/services/ct-payment.service';
-import { Payment } from '@commercetools/platform-sdk';
-import { Transaction } from '@commercetools/platform-sdk/dist/declarations/src';
+import { mockGetPaymentResult, mockUpdatePaymentResult } from './utils/mock-payment-data';
 
 describe('operation.service', () => {
   const opts: OperationServiceOptions = {
@@ -62,57 +61,6 @@ describe('operation.service', () => {
           },
         ],
       },
-    };
-
-    const mockGetPaymentResult: Payment = {
-      id: '123456',
-      version: 1,
-      amountPlanned: {
-        type: 'centPrecision',
-        currencyCode: 'GBP',
-        centAmount: 120000,
-        fractionDigits: 2,
-      },
-      paymentMethodInfo: {
-        method: 'Debit Card',
-        name: { 'en-US': 'Debit Card', 'en-GB': 'Debit Card' },
-      },
-      paymentStatus: { interfaceText: 'Paid' },
-      transactions: [],
-      interfaceInteractions: [],
-      createdAt: '2024-02-13T00:00:00.000Z',
-      lastModifiedAt: '2024-02-13T00:00:00.000Z',
-    };
-    const mockCancelPaymentTransaction: Transaction = {
-      id: 'dummy-transaction-id',
-      timestamp: '2024-02-13T00:00:00.000Z',
-      type: 'CancelAuthorization',
-      amount: {
-        type: 'centPrecision',
-        centAmount: 120000,
-        currencyCode: 'GBP',
-        fractionDigits: 2,
-      },
-      state: 'Initial',
-    };
-    const mockUpdatePaymentResult: Payment = {
-      id: '123456',
-      version: 1,
-      amountPlanned: {
-        type: 'centPrecision',
-        currencyCode: 'GBP',
-        centAmount: 120000,
-        fractionDigits: 2,
-      },
-      paymentMethodInfo: {
-        method: 'Debit Card',
-        name: { 'en-US': 'Debit Card', 'en-GB': 'Debit Card' },
-      },
-      paymentStatus: { interfaceText: 'Paid' },
-      transactions: [mockCancelPaymentTransaction],
-      interfaceInteractions: [],
-      createdAt: '2024-02-13T00:00:00.000Z',
-      lastModifiedAt: '2024-02-13T00:00:00.000Z',
     };
 
     sinon.stub(DefaultPaymentService.prototype, 'getPayment').callsFake(async () => {
