+++ processor/test/utils/mock-payment-data.ts
@@ -1,0 +1,54 @@
+import { Payment } from '@commercetools/platform-sdk';
+import { Transaction } from '@commercetools/platform-sdk/dist/declarations/src';
+export const mockGetPaymentResult: Payment = {
+  id: '123456',
+  version: 1,
+  amountPlanned: {
+    type: 'centPrecision',
+    currencyCode: 'GBP',
+    centAmount: 120000,
+    fractionDigits: 2,
+  },
+  paymentMethodInfo: {
+    method: 'Debit Card',
+    name: { 'en-US': 'Debit Card', 'en-GB': 'Debit Card' },
+  },
+  paymentStatus: { interfaceText: 'Paid' },
+  transactions: [],
+  interfaceInteractions: [],
+  createdAt: '2024-02-13T00:00:00.000Z',
+  lastModifiedAt: '2024-02-13T00:00:00.000Z',
+};
+
+const mockCancelPaymentTransaction: Transaction = {
+  id: 'dummy-transaction-id',
+  timestamp: '2024-02-13T00:00:00.000Z',
+  type: 'CancelAuthorization',
+  amount: {
+    type: 'centPrecision',
+    centAmount: 120000,
+    currencyCode: 'GBP',
+    fractionDigits: 2,
+  },
+  state: 'Initial',
+};
+
+export const mockUpdatePaymentResult: Payment = {
+  id: '123456',
+  version: 1,
+  amountPlanned: {
+    type: 'centPrecision',
+    currencyCode: 'GBP',
+    centAmount: 120000,
+    fractionDigits: 2,
+  },
+  paymentMethodInfo: {
+    method: 'Debit Card',
+    name: { 'en-US': 'Debit Card', 'en-GB': 'Debit Card' },
+  },
+  paymentStatus: { interfaceText: 'Paid' },
+  transactions: [mockCancelPaymentTransaction],
+  interfaceInteractions: [],
+  createdAt: '2024-02-13T00:00:00.000Z',
+  lastModifiedAt: '2024-02-13T00:00:00.000Z',
+};
